requirements.txt
streamlit
pandas
plotly
requests
python-docx
yfinance
numpy
